pyproject.toml
[tool.poetry]
name = "meerge-africa"
version = "0.1.0"
description = "The Codebase For http://www.meergeafrica.com/"
authors = ["John David <davemcsavvii@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = ">=4.2,<5.1"
wagtail = ">=6.2,<6.3"
gunicorn = "^22.0.0"
django-htmx = "^1.18.0"
djangorestframework = "^3.15.2"
django-allauth = "^0.63.6"
django-ninja = "^1.2.2"
django-seed = "^0.3.1"
django-extensions = "^3.2.3"
poetry-plugin-export = "^1.8.0"
pyjwt = "^2.8.0"
django-formtools = "^2.5.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
django-cities-light = "^3.10.1"
django-compressor = "^4.5.1"
django-browser-reload = "^1.13.0"
cryptography = "^43.0.0"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
django-unfold = "^0.38.0"
django-phonenumbers = "^1.0.1"
pyopenssl = "^24.2.1"
django-rest-swagger = "^2.2.0"
pyyaml = "^6.0.2"
uritemplate = "^4.1.1"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.0"
ruff = "^0.5.5"
ipython = "^8.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 80
indent-width = 4
target-version = "py310"
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".venv",
    ".vscode",
    ".ruff_cache",
    ".mypy_cache",
    ".pytest_cache",
    "*.ipynb"
]

[tool.mypy]
ignore_missing_imports = true